requirements.txt
azure-functions==1.20.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
azure-cognitiveservices-language-textanalytics==0.2.0
openai>=1.55.3
python-dotenv==1.0.0
requests==2.31.0
python-docx==1.1.0
webvtt-py==0.4.6
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0
msal==1.24.0
